requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
idna==2.8
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.2.2
pytz==2018.9
requests==2.21.0
six==1.12.0
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
